move heads (events without descendants) from poset to node
