+++ src/gossip/store_heads.go
@@ -1,4 +1,4 @@
-package posposet
+package gossip
 
 import (
 	"github.com/Fantom-foundation/go-lachesis/src/hash"
@@ -7,7 +7,7 @@ import (
 func (s *Store) EraseHead(id hash.Event) {
 	key := id.Bytes()
 
-	if err := s.epochTable.Heads.Delete(key); err != nil {
+	if err := s.table.Heads.Delete(key); err != nil {
 		s.Fatal(err)
 	}
 }
@@ -15,7 +15,7 @@ func (s *Store) EraseHead(id hash.Event) {
 func (s *Store) AddHead(id hash.Event) {
 	key := id.Bytes()
 
-	if err := s.epochTable.Heads.Put(key, []byte{}); err != nil {
+	if err := s.table.Heads.Put(key, []byte{}); err != nil {
 		s.Fatal(err)
 	}
 }
@@ -23,17 +23,18 @@ func (s *Store) AddHead(id hash.Event) {
 func (s *Store) IsHead(id hash.Event) bool {
 	key := id.Bytes()
 
-	ok, err := s.epochTable.Heads.Has(key)
+	ok, err := s.table.Heads.Has(key)
 	if err != nil {
 		s.Fatal(err)
 	}
 	return ok
 }
 
+// GetHeads returns all the events with no descendants
 func (s *Store) GetHeads() hash.Events {
 	prefix := []byte{}
 	res := []hash.Event{}
-	err := s.epochTable.Heads.ForEach(prefix, func(key, _ []byte) bool {
+	err := s.table.Heads.ForEach(prefix, func(key, _ []byte) bool {
 		res = append(res, hash.BytesToEvent(key))
 		return true
 	})
