+++ src/posposet/store.go
@@ -29,7 +29,6 @@ type Store struct {
 	}
 
 	epochTable struct {
-		Heads       kvdb.Database `table:"heads_"`
 		Roots       kvdb.Database `table:"roots_"`
 		VectorIndex kvdb.Database `table:"vectors_"`
 	}
